+++ src/main/java/mediathek/config/Konstanten.java
@@ -47,7 +47,10 @@ public class Konstanten {
 
     public static final byte MAX_DOWNLOAD_RESTARTS = 2;
     public static final byte MAX_EXTERNAL_STARTS = 3;
-    public static final byte CONTINUE_DOWNLOAD = 60; //seconds
+    /**
+     * Default time for automatic continuation of existing downloads.
+     */
+    public static final byte DOWNLOAD_CONTINUATION_DEFAULT_TIME = 60; //seconds
     public static final byte DOWNLOAD_ERROR_DISPLAY_DURATION = 60;
 
     // MediathekView URLs